Rework generative test support to fully control side effects
Previously, steps were treated as no-op generators that are glued together with
bind. This is not what test.check is expecting, and sometimes it will outright
skip some of them. This is problematic if you're using said steps to do
something with an external system.

Now steps are executed explicitly. This means that generators can't rely on the
results of steps (the two are logically evaluated in parallel) and all
generators have to happen in a block together.